pyproject.toml
[project]
name = "fin-research-agents"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
"langgraph>=0.2.0",
"langchain>=0.3.0",
"langchain-community>=0.3.0",
"anthropic>=0.39.0",
"groq>=0.9.0",
"chromadb>=0.5.0",
"sentence-transformers>=3.0.0",
"yfinance>=0.2.40",
"requests>=2.31.0",
"python-dotenv>=1.0.0",
"pydantic>=2.7.0",
"mcp>=1.0.0",
"fastmcp>=0.2.0",
"typer>=0.12.0",
"rich>=13.7.0",
]
[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.5", "ipykernel"]

[tool.pytest.ini_options]
pythonpath = ["src"]